requirements.txt
asgiref==3.5.0
AWSIoTPythonSDK==1.4.9
bidict==0.21.4
Django==4.0.2
djangorestframework==3.13.1
dnspython==2.2.0
eventlet==0.33.0
greenlet==1.1.2
gunicorn==20.1.0
portalocker==2.3.2
pymongo==4.0.1
pypiwin32==223
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
pywin32==303
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
